refactored create post feature
