+++ Client/src/components/PrivateRoute.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import {Redirect, Route} from 'react-router-dom'
-import {useUserContext} from '../src/AuthContext'
+import {useUserContext} from '../AuthContext'
 
 const PrivateRoute = ({component: Component, ...rest}) => {
     const {user} = useUserContext();
